세션 -> 토큰

+++ ReciPT/src/main/java/samdasu/recipt/domain/controller/UserApiController.java
@@ -7,10 +7,9 @@ import lombok.Data;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 import samdasu.recipt.domain.controller.dto.Heart.RecipeHeartDto;
@@ -19,10 +18,12 @@ import samdasu.recipt.domain.controller.dto.Register.RegisterResponseDto;
 import samdasu.recipt.domain.controller.dto.User.UserResponseDto;
 import samdasu.recipt.domain.controller.dto.User.UserSignUpDto;
 import samdasu.recipt.domain.controller.dto.User.UserUpdateRequestDto;
+import samdasu.recipt.domain.entity.Heart;
 import samdasu.recipt.domain.entity.RegisterRecipe;
 import samdasu.recipt.domain.entity.Review;
 import samdasu.recipt.domain.entity.User;
 import samdasu.recipt.domain.service.*;
+import samdasu.recipt.security.config.auth.PrincipalDetails;
 
 import javax.validation.Valid;
 import java.io.IOException;
@@ -30,6 +31,10 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
+
+import static org.springframework.http.HttpStatus.FORBIDDEN;
+import static org.springframework.http.HttpStatus.OK;
+import static org.springframework.http.ResponseEntity.status;
 
 @Slf4j
 @RestController
@@ -39,8 +44,8 @@ public class UserApiController {
     private final UserService userService;
     private final ProfileService profileService;
     private final RegisterRecipeService registerRecipeService;
-
     private final ReviewService reviewService;
+    private final HeartService heartService;
 
     @PostMapping("/signup")
     public Result1 saveUser(@Valid UserSignUpDto userSignUpDto, @RequestParam(value = "profile") MultipartFile file) throws IOException {
@@ -53,7 +58,7 @@ public class UserApiController {
 
         byte[] downloadImage = profileService.downloadImage(findUser.getProfile().getProfileId()); //프로필 사진
 
-        return new Result1(1, new UserResponseDto(findUser), ResponseEntity.status(HttpStatus.OK)
+        return new Result1(1, new UserResponseDto(findUser), status(OK)
                 .contentType(MediaType.valueOf("image/png"))
                 .body(downloadImage));
     }
@@ -62,8 +67,9 @@ public class UserApiController {
      * 프로필 보기
      */
     @GetMapping("/user")
-    public Result1 userInfo(@AuthenticationPrincipal UserResponseDto userResponseDto) throws JsonProcessingException {
-        User findUser = userService.findById(userResponseDto.getUserId());
+    public Result1 userInfo(Authentication authentication) throws JsonProcessingException {
+        PrincipalDetails principal = (PrincipalDetails) authentication.getPrincipal();
+        User findUser = userService.findById(principal.getUser().getUserId());
         byte[] downloadImage = profileService.downloadImage(findUser.getProfile().getProfileId()); //프로필 사진
 
         log.info("user.getUsername() = {}", findUser.getUsername());
@@ -85,18 +91,19 @@ public class UserApiController {
         headers.setContentType(MediaType.APPLICATION_JSON);
         headers.setContentLength(jsonData.length);
 
-        return new Result1(1, responseDto, new ResponseEntity<>(jsonData, headers, HttpStatus.OK));
+        return new Result1(1, responseDto, new ResponseEntity<>(jsonData, headers, OK));
     }
 
     /**
      * 프로필 수정
      */
     @PostMapping("/user/edit")
-    public Result1 updateUser(@AuthenticationPrincipal UserResponseDto userResponseDto,
+    public Result1 updateUser(Authentication authentication,
                               @Valid UserUpdateRequestDto request) throws JsonProcessingException {
-        userService.update(userResponseDto.getUserId(), request);
+        PrincipalDetails principal = (PrincipalDetails) authentication.getPrincipal();
+        Long updateUserId = userService.update(principal.getUser().getUserId(), request);
 
-        User findUser = userService.findById(userResponseDto.getUserId());
+        User findUser = userService.findById(updateUserId);
         byte[] downloadImage = profileService.downloadImage(findUser.getProfile().getProfileId()); //프로필 사진
 
         log.info("user.getPassword() = {}", findUser.getPassword());
@@ -115,18 +122,17 @@ public class UserApiController {
         headers.setContentType(MediaType.APPLICATION_JSON);
         headers.setContentLength(jsonData.length);
 
-        return new Result1(1, responseDto, new ResponseEntity<>(jsonData, headers, HttpStatus.OK));
+        return new Result1(1, responseDto, new ResponseEntity<>(jsonData, headers, OK));
     }
 
     /**
      * 유저가 누른 좋아요 보기
      */
 
-    private final HeartService heartService;
-
     @GetMapping("/user/like")
-    public Result2 searchLikeInfo(@AuthenticationPrincipal UserResponseDto userResponseDto) {
-        User findUser = userService.findById(userResponseDto.getUserId());
+    public Result2 searchLikeInfo(Authentication authentication) {
+        PrincipalDetails principal = (PrincipalDetails) authentication.getPrincipal();
+        User findUser = userService.findById(principal.getUser().getUserId());
 
 //        RecipeHeartDto recipeHeartDto = RecipeHeartDto.createRecipeHeartDto(findUser.getUserId(), 3L, "된장 부대찌개", "찌개", "다시마 1g, 두부 10g, 떡국 떡 10g, 스팸(마일드) 10g, 다진 마늘 5g, 무 20g, 김치 15g, 소시지 10g(1/2개), 우민찌 5g(1작은술), 양파 5g, 저염된장 15g(1큰술), 베이컨 5g, 대파 5g, 청양고추 5g, 홍고추 1g");
 //        RecipeHeartDto recipeHeartDto1 = RecipeHeartDto.createRecipeHeartDto(findUser.getUserId(), 10L, "구운채소", "채소", "가지 20g(3cm), 호박 50g(1/3개), 새송이버섯 15g(3개), 양파 15g(1/8개), 발사믹크레마 15g(1큰술), 빨강 파프리카 3g(3×1cm), 노랑 파프리카 3g(3×1cm), 청피망 3g(3×1cm), 올리브유 약간, 저염간장 5g(1작은술), 식초 15g(1큰술), 설탕 10g(1큰술), 레몬즙 5g(1작은술)");
@@ -164,8 +170,9 @@ public class UserApiController {
      * 유저가 등록한 레시피 보기
      */
     @GetMapping("/user/register")
-    public Result1 searchRegisterInfo(@AuthenticationPrincipal UserResponseDto userResponseDto) {
-        User findUser = userService.findById(userResponseDto.getUserId());
+    public Result1 searchRegisterInfo(Authentication authentication) {
+        PrincipalDetails principal = (PrincipalDetails) authentication.getPrincipal();
+        User findUser = userService.findById(principal.getUser().getUserId());
         UserResponseDto responseDto = new UserResponseDto(findUser);
         byte[] downloadImage = profileService.downloadImage(findUser.getProfile().getProfileId()); //프로필 사진
 
@@ -177,37 +184,54 @@ public class UserApiController {
                 .map(RegisterResponseDto::new)
                 .collect(Collectors.toList());
 
-        return new Result1(collect.size(), responseDto, ResponseEntity.status(HttpStatus.OK)
+        return new Result1(collect.size(), responseDto, status(OK)
                 .contentType(MediaType.valueOf("image/png"))
                 .body(downloadImage));
     }
 
     /**
      * 등록된 레시피 삭제
-     * - 삭제 후 redirect:/user/register
      */
     @PostMapping("/user/delete")
-    public void deleteRecipe(@AuthenticationPrincipal UserResponseDto userResponseDto
+    public ResponseEntity<String> deleteRecipe(Authentication authentication
             , @RequestParam(value = "registerRecipeId") Long registerRecipeId) {
-        registerRecipeService.deleteRegisterRecipe(registerRecipeId);
+        boolean isDelete = false;
+        PrincipalDetails principal = (PrincipalDetails) authentication.getPrincipal();
+        User findUser = userService.findById(principal.getUser().getUserId());
+        List<RegisterRecipe> registerRecipes = findUser.getRegisterRecipes();
+
+        for (RegisterRecipe registerRecipe : registerRecipes) {
+            if (registerRecipe.getRegisterId().equals(registerRecipeId)) {
+                isDelete = true;
+                List<Heart> hearts = registerRecipe.getHearts();
+                for (Heart heart : hearts) {
+                    RegisterHeartDto registerHeartDto = RegisterHeartDto.createRegisterHeartDto(heart);
+                    heartService.deleteRegisterRecipeHeart(registerHeartDto);
+                }
+                registerRecipeService.deleteRegisterRecipe(registerRecipeId);
+                return ResponseEntity.status(OK).body("레시피 삭제 완료!");
+            }
+        }
+        return status(FORBIDDEN).body("삭제 권한이 없습니다!");
     }
 
     /**
      * 유저가 작성한 리뷰 보기
      */
     @GetMapping("/user/review")
-    public Result1 searchReviewInfo(@AuthenticationPrincipal UserResponseDto userResponseDto) {
-        User findUser = userService.findById(userResponseDto.getUserId());
+    public Result1 searchReviewInfo(Authentication authentication) {
+        PrincipalDetails principal = (PrincipalDetails) authentication.getPrincipal();
+        User findUser = userService.findById(principal.getUser().getUserId());
         UserResponseDto responseDto = new UserResponseDto(findUser);
         byte[] downloadImage = profileService.downloadImage(findUser.getProfile().getProfileId()); //프로필 사진
 
-        List<Review> reviews = reviewService.findReviewByWriter(userResponseDto.getUsername());
+        List<Review> reviews = reviewService.findReviewByWriter(findUser.getUsername());
 
         log.info("user.getUsername() = {}", findUser.getUsername());
         log.info("user.getReviews.getComment() = {}", findUser.getReviews().stream()
                 .map(Review::getComment).collect(Collectors.toList()));
 
-        return new Result1(reviews.size(), responseDto, ResponseEntity.status(HttpStatus.OK)
+        return new Result1(reviews.size(), responseDto, status(OK)
                 .contentType(MediaType.valueOf("image/png"))
                 .body(downloadImage));
     }
